chore: renaming + tidying up things, removing bind
