+++ src/pages/Maps/Content/View/Sprites.tsx
