+++ src/pages/Maps/Content/index.ts
@@ -1,0 +1,4 @@
+import { MapView } from './View'
+import { Controls } from './Controls'
+
+export { MapView, Controls }
